fix(ui): avoid unnecessary FluxTable[] parsing
Also consolidates the `VisSwitcher` and `RefreshingViewSwitcher`
components into a single `ViewSwitcher` component.

Closes #13852
